Refactor data source classes to fix import issues (#1723)
* Refactor data source classes to fix import issues


* make format and lint


* remove unneded __init__ files
