+++ sdk/python/feast/__init__.py
@@ -2,9 +2,9 @@ import logging
 
 from pkg_resources import DistributionNotFound, get_distribution
 
-from feast.infra.offline_stores.bigquery import BigQuerySource
-from feast.infra.offline_stores.file import FileSource
-from feast.infra.offline_stores.redshift import RedshiftSource
+from feast.infra.offline_stores.bigquery_source import BigQuerySource
+from feast.infra.offline_stores.file_source import FileSource
+from feast.infra.offline_stores.redshift_source import RedshiftSource
 
 from .client import Client
 from .data_source import KafkaSource, KinesisSource, SourceType
@@ -29,12 +29,10 @@ except DistributionNotFound:
     pass
 
 __all__ = [
-    "BigQuerySource",
     "Client",
     "Entity",
     "KafkaSource",
     "KinesisSource",
-    "RedshiftSource",
     "Feature",
     "FeatureStore",
     "FeatureTable",
@@ -42,5 +40,7 @@ __all__ = [
     "RepoConfig",
     "SourceType",
     "ValueType",
+    "BigQuerySource",
     "FileSource",
+    "RedshiftSource",
 ]
