+++ sdk/python/usage_tests/test_usage.py
@@ -187,7 +187,7 @@ def test_exception_usage_off():
     assert rows.total_rows == 0
 
 
-@retry(wait=wait_exponential(multiplier=1, min=1, max=10), stop=stop_after_attempt(5))
+@retry(wait=wait_exponential(multiplier=1, min=1, max=10), stop=stop_after_attempt(7))
 def ensure_bigquery_usage_id_with_retry(usage_id):
     rows = read_bigquery_usage_id(usage_id)
     if rows.total_rows != 1:
